% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSpecies.R
\name{cleanSpecies}
\alias{cleanSpecies}
\title{Helper function to clean taxonomy.}
\usage{
cleanSpecies(species, verbose = TRUE, db = "itis")
}
\arguments{
\item{species}{A vector of species to check}

\item{verbose}{logical. Print stuff? Default to TRUE}

\item{db}{Which db should be used 'itis' is the default}
}
\value{
data.frame of original and corrected species.
}
\description{
This correct misspellings, synonyms and flags not recognizes species.
}
\examples{
#not run
species <- c("Osmia rufa", "Osmia bicornis", "Osmia ruffa",
           "Osmia wikifluqie", "watermelon pie", "Osmia sp.")
cleanSpecies(species)
}
