% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Count the number of rows in a query.}
\usage{
count(resource, conditions = NULL, distinct = NULL)
}
\arguments{
\item{resource}{The resource to query.}

\item{conditions}{A set of conditions to constraint the query on the server side. Use the function \code{buildCondition()} to create the conditions.0}

\item{distinct}{Provides a distict criteria. (Not yet implemented)}
}
\value{
Returns the total count of rows in a query.
}
\description{
Returns the total count of rows in a query.
}
\examples{
cnx <- connect('http://www.traitbase.info', 'demo', '1234')
off <- resource(cnx, 'species')

count(off)
count(off, conditions=buildCondition('species', '!=', 'Bombus')  )
}
