% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDatasetById.R
\name{deleteDatasetById}
\alias{deleteDatasetById}
\title{Deelet a dataset from Traitbase.}
\usage{
deleteDatasetById(cnx, datasetId)
}
\arguments{
\item{cnx}{The connection to traitbase service.}

\item{datasetId}{The ID of the dataset to delete}
}
\value{
Returns a list of errors if any or OK.
}
\description{
Deelet a dataset from Traitbase.
}
\examples{
\donttest{
datasetId <- 'ds1234'
cnx <- connect('http://www.traitbase.info')
errors <- importDataset(cnx, datasetId)
}
}
