% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseErrors.R
\name{parseErrors}
\alias{parseErrors}
\title{Helper function to parse errors fron validateDataset().}
\usage{
parseErrors(errors)
}
\arguments{
\item{errors}{Output of validateDataset()}
}
\value{
list of errors (err) and lines where encountered (tax).
}
\description{
Transform the hard to read output of validateDataset() in
a usefull format.
}
\examples{
#not run

}
