% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryRaw.R
\name{queryRaw}
\alias{queryRaw}
\title{Makes a raw query}
\usage{
queryRaw(resource, query)
}
\arguments{
\item{resource}{The resource to query.}

\item{query}{Raw URL query. See doc here: https://github.com/wprl/baucis/wiki/Query-String-Parameters}
}
\value{
Returns a dataframe with the returned data.
}
\description{
Executes a query allowing pasing a direct Hivepod/Baucis query using JSON
syntax as defined here:
https://github.com/wprl/baucis/wiki/Query-String-Parameters
}
\examples{
\donttest{
cnx <- connect('http://www.traitbase.info', 'demo', '1234')
spe <- resource(cnx, 'species')

queryRaw(spe, '?skip=2&limit=3')
}
}
