% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource}
\alias{resource}
\title{Creates a resource.}
\usage{
resource(connection, resourceName)
}
\arguments{
\item{connection}{The connection created with the \code{connect()} function.}

\item{resourceName}{The resource name used in Hivepod. Note that is uses to be a plural name.}
}
\value{
\code{resource()} returns a resource object for a Hivepod backend.
This resource object can be used to make queries, count and manipulate data in the backend.
}
\description{
Setups and returns a resource available in a Hivepod backend.
}
\examples{
\donttest{
cnx <- connect('http://www.traitbase.info')
exo <- resource(cnx, 'species')
whe <- resource(cnx, 'schemas')
}
}
