% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDataset.R
\name{validateDataset}
\alias{validateDataset}
\title{Validates a dataset prior an import operation.}
\usage{
validateDataset(cnx, csvData)
}
\arguments{
\item{cnx}{The connection to traitbase service.}

\item{csvData}{The data in cvs format (string)}
}
\value{
Returns a list of errors if any. valid=(true|false), imported=false
}
\description{
Validates a dataset prior an import operation.
}
\examples{
cnx <- connect("http://www.traitbase.info", "demo", "1234")
fpath <- system.file("extdata", "sample-error.csv", package="traitbaser")
csvData <- readLines(fpath)
errors <- validateDataset(cnx, csvData)
}

